refactor role manage (#24146)
